requirements.txt
python-telegram-bot==20.4
google-api-python-client==2.95.0
black==23.7.0
isort==5.12.0
pre-commit==3.3.3
